common: move json_to_blinded_path into its own file.
It's fairly obscure, and let's not make everyone link it.
