+++ common/json_parse.h
@@ -114,10 +114,6 @@ bool json_to_channel_id(const char *buffer, const jsmntok_t *tok,
 bool json_to_coin_mvt_tag(const char *buffer, const jsmntok_t *tok,
 			  enum mvt_tag *tag);
 
-/* Extract reply path from this JSON */
-struct blinded_path *
-json_to_blinded_path(const tal_t *ctx, const char *buffer, const jsmntok_t *tok);
-
 bool json_tok_channel_id(const char *buffer, const jsmntok_t *tok,
 			 struct channel_id *cid);
